credential-broker-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
aiofiles==24.1.0
pydantic==2.11.7
asyncpg==0.29.0
sqlalchemy==1.4.27
pydantic-settings==2.10.1
python-dotenv==1.1.1
cryptography==45.0.6
